add exception for unshare(2) failures
Add custom exception and exit code to allow higher level logic
calling mock to retry when unshare(2) fails.

+++ py/mock/plugins/selinux.py
@@ -68,11 +68,12 @@ class SELinux(object):
 
     decorate(traceLog())
     def _selinuxAtExit(self):
-        try:
-            os.unlink(self.filesystems)
-        except OSError, e:
-            getLog().warning("unable to delete selinux filesystems (%s): %s" % (self.filesystems, e))
-            pass
+        if os.path.exists(self.filesystems):
+            try:
+                os.unlink(self.filesystems)
+            except OSError, e:
+                getLog().warning("unable to delete selinux filesystems (%s): %s" % (self.filesystems, e))
+                pass
 
     decorate(traceLog())
     def _selinuxPreYumHook(self):
